+++ tests/spec/libsass_closed_issues/issue_1257.rs
@@ -1,7 +1,7 @@
 //! Tests auto-converted from "sass-spec/spec/libsass-closed-issues/issue_1257.hrx"
 
 #[test]
-#[ignore] // wrong result
+#[ignore] // unexepected error
 fn test() {
     assert_eq!(
         crate::rsass(