+++ CHANGELOG.md
@@ -22,7 +22,8 @@ project adheres to
 * Improved parameter handling and returned values of the supported
   selector functions.  PR #103.
 * Implement `meta.module_variables` and `meta.module_functions`.
-* Improved parameter checking for `color.hwb` function.
+* Improved parameter checking for `hwb`, `alpha`, and `invert`
+  functions in `sass:color` module.
 
 
 ## Release 0.20.0 - 2021-03-25